🔧 Fix table of content

+++ projects/notedown-ast/src/nodes/elements/header.rs
@@ -2,13 +2,13 @@ use crate::nodes::*;
 
 #[derive(Clone, Debug, Eq, PartialEq, Hash)]
 pub struct Header {
-    pub level: usize,
+    pub level: u8,
     pub children: Vec<ASTNode>,
 }
 
 impl Display for Header {
     fn fmt(&self, f: &mut Formatter) -> fmt::Result {
-        write!(f, "{}", "#".repeat(self.level),)?;
+        write!(f, "{}", "#".repeat(self.level as usize),)?;
         for term in &self.children {
             write!(f, "{}", term)?
         }
